Run tests and lint, including simple integration test, in CI (#5)
Previously, CI was only to auto-deploy the demo app to
https://knit-demo.connect.build.

But it should also run tests and perform linting, to make sure a PR
doesn't break stuff, and it should push to the BSR when the protos
change. So this adds that. It also adds a simple integration test that
starts up the server, a knit gateway, and then runs the `ts/index.ts`
script to hit the gateway and make sure a query works.